code cleanup

+++ meja-generic/src/main/java/com/dua3/meja/model/generic/GenericCellStyle.java
@@ -257,7 +257,7 @@ public class GenericCellStyle implements CellStyle {
     @Override
     public void setRotation(short angle) {
         LangUtil.check(angle>=-90 && angle <=90, "angle must be in range [-90, 90]: %d", angle);
-        this.rotation = (short) angle;
+        this.rotation = angle;
     }
 
     @Override
